Added some graphics tools
While this. version produces well-formed cubes, some nodes seem to get trapped in a given fractality, an no adjacent cube can be divided. In the next version, we will not only reorient the ExpandManyRandomly subroutine to favor the. division of "mixed fractality" cubes (since these serve to expand an already existing region of higher fractality at the expense of regions of lower fractality) as opposed to those where the fractality is niform, which create a new region of higher fractality where none had existed previously. The two types are easily distinguishable from the relevant coord-path (tree of paths constituting the edges of the cube), in that mixed-fractality cubes always have some edges that are divided, whereas in the uniform-scenario divisions, the corners of the cube are all nearest neighbors (i.e. the dge paths are all of length two).