Last version before advancing to "edge" topology
Up to and including this file, I've assumed that "edge-ness" can be ascertained from neighbor count alone, and in 2-d, that is indeed sufficient. But in 3, when trying to divide a cube of which 2 abutting faces are subdivided already, the edge that they have in common has 6 neighbors, as does a node that is not an edge at all. In that case, we need to instead categorize a node by number of adjacent cubes which have an "opposite node" of the respective cube obtained by simple intersection (and maybe neighbor count as well, or maybe the former is to be assessed only when the latter is insufficient to mark the node as being on an edge.