Added NRegions and RegionHist
Added specifiers to account for the fact that a node may have 2*D neighbors (the maximum allowable amount) and still not an edge or subsurface of one form or another, as is the case with, say, two Rubik's cubes that are aligned like two steps on a staircase so that they have only one linear segment in common. The nodes along
that common segment all have 2D (i.e. 6) neighbors, and yet only "half" the space around any of those nodes (each of which will eventually be the intersection of 2**D cubes) is only half occupied by divided cubes. That means we have to make. use of more than just neighbor count in order to. identify edges or faces or other subsurfaces.