Finally something that seems to work
Transitioned from testing on len(node.Neighbors) != 2*self.NDim to testing on node.NRegions != 2**self.NDim; the latter (unlike the former) works for dimensions higher than 2. In particular, the former fails for configurations consisting of a grid where every cube had the same fractality save for a linear series of adjacent cubes of one lower level of fractality. Note that the number of neighbors for each node along the edges of this tunnel (i.e. not the faces, but the actual corner edges) is 2*D (i.e. fully packed), so that the earlier test thought it was not an edge at all.