Latest set of simulation runs
This version was used to generate many different cube-universes, and in each of them the starting nodes all got subdivided with little difficulty. I.e. there was no configuration among all these attempts in which a surrounding set of cubes caused a given cube to be indivisible. That helps assure us that eventually all the created subcubes will themselves be divisible, allowing us to invoke Fick's Laws of Diffusion in which the heat equation (and by extension, the wave and Klein-Gordon equations) exist with a well defined diffusion constant and dimensionality of 3 (or 2, as with earlier runs) as opposed to the freaky Hasdorff dimensional spaces we kept getting earlier with. other kinds of fractually dividing cubes and simplices.